refactor: profile folder structer optimization
