+++ src/pages/profile/Profile.tsx
@@ -1,38 +1,39 @@
-import { Box, Chip, Stack } from '@mui/material';
+import { Box, Button, Chip, Stack } from '@mui/material';
 import { motion } from 'motion/react';
-import { useState } from 'react';
 
 import { Loading, Title } from '../../components';
-import { useAppSelector } from '../../redux/hooks';
+import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import { getCustomer } from '../../redux/selectors';
-import AuthLayout from './AuthLayout';
-import GuestLayout from './GuestLayout';
+import { setIsEditProfile } from '../../redux/slices/authSlice';
+import AuthLayout from './auth-layout/AuthLayout';
+import GuestLayout from './guest-layout/GuestLayout';
 
 const Profile = () => {
-  const { customer, isLoading } = useAppSelector(getCustomer);
-  const [isEditMode, setIsEditMode] = useState(false);
+  const { customer, isLoading, isEditProfile } = useAppSelector(getCustomer);
+  const dispatch = useAppDispatch();
 
   return (
     <Box animate={{ scale: 1 }} component={motion.div} initial={{ scale: 0 }}>
       <Stack alignItems={'center'} direction={'row'}>
         <Title variant="main">Profile</Title>
         <Chip
-          label={isEditMode ? 'Edit mode' : 'View mode'}
-          sx={{ bgcolor: isEditMode ? 'primary.main' : 'disabled', m: 4 }}
+          label={isEditProfile ? 'Edit mode' : 'View mode'}
+          sx={{ bgcolor: isEditProfile ? 'primary.main' : 'disabled', m: 4 }}
         />
+        <Button
+          disabled={!customer}
+          onClick={() => {
+            dispatch(setIsEditProfile(!isEditProfile));
+          }}
+          sx={{ ml: 'auto' }}
+          variant={isEditProfile ? 'contained' : 'outlined'}
+        >
+          Edit
+        </Button>
       </Stack>
-      <>
-        {isLoading ? <Loading /> : null}
-        {customer ? (
-          <AuthLayout
-            customer={customer}
-            isEditMode={isEditMode}
-            setIsEditMode={setIsEditMode}
-          />
-        ) : (
-          <GuestLayout />
-        )}
-      </>
+
+      {isLoading ? <Loading /> : null}
+      {customer ? <AuthLayout /> : <GuestLayout />}
     </Box>
   );
 };
