+++ src/pages/profile/guest-layout/GuestLayout.tsx
@@ -2,8 +2,8 @@ import { Button, Typography } from '@mui/material';
 import { useEffect, useState } from 'react';
 import { Link, useNavigate } from 'react-router';
 
-import { Title } from '../../components';
-import CONSTANTS from '../../utils/CONSTANTS';
+import { Title } from '../../../components';
+import CONSTANTS from '../../../utils/CONSTANTS';
 
 const GuestLayout = () => {
   const navigate = useNavigate();
